Add unit test and code refactor
